requirements.txt
certifi==2022.6.15
chardet==5.0.0
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
h5py==3.7.0
htmlmin==0.1.12
itsdangerous==2.1.2
jsonschema==4.9.1
keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.1
numpydoc==1.4.0
pandas==1.4.3
pandas-profiling==3.2.0
Pillow==9.2.0
requests==2.28.1
requests-oauthlib==1.3.1
Rtree==1.0.0
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
sklearn==0.0
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
wincertstore==0.2